subscriptions controller plus testing of them

+++ src/controllers/video.controller.js
@@ -115,9 +115,54 @@ const getVideoById = asyncHandler(async (req, res) => {
 })
 
 const updateVideo = asyncHandler(async (req, res) => {
-    const { videoId } = req.params
-    //TODO: update video details like title, description, thumbnail
-    // next task
+try {
+        const { videoId } = req.params
+        // below expression checks if videoId is a valid object id
+        if (!mongoose.Types.ObjectId.isValid(videoId)) {
+            return res.status(400).json({ error: 'Invalid videoId' });
+        }
+        //TODO: update video details like title, description, thumbnail
+        const thumbnailLocalPath = req.file?.path;
+        if(!thumbnailLocalPath){
+            throw new ApiError(400,"thumbnail not found")
+        }
+        const user = req.user;
+        if(!user){
+            throw new ApiError(400,"user not found!")
+        }
+        const video = await Video.findById(videoId);
+        if(video.owner.toString() !== user._id.toString()){
+            throw new ApiError(400,`user don't have permission to update this video`)
+        }  
+        const thumbnail = await uploadOnCloudinary(thumbnailLocalPath)
+        const {title,description} = req.body;
+        const updatedVideo = await Video.findByIdAndUpdate(
+             videoId,
+             {
+                $set : {
+                    thumbnail :  thumbnail.url,
+                    title,
+                    description
+                }
+             },
+             {
+                new : true
+             }
+        )
+        if(!updatedVideo){
+            throw new ApiError(500,"Error while updating the video")
+        }
+
+        res.status(200).json(
+            new ApiResponse(200,updatedVideo,"Video successfully update")
+        )
+} catch (error) {
+    if(error instanceof ApiError){
+        throw error
+    }else{
+        throw new ApiError(500,error?.message || "Something went wrong while updating the video")
+    }
+}
 
 })
 
@@ -131,6 +176,16 @@ const deleteVideo = asyncHandler(async (req, res) => {
         // commenting the below code because coludinary doesn't delete by url instead by public id
         //const {videoFile , thumbnail} = await Video.findById(videoId);
         */
+       // checking if the user is owner of this video
+        const user = req.user;
+        if(!user){
+            throw new ApiError(400,"user not found!")
+        }
+        const video = await Video.findById(videoId);
+        if(video.owner.toString() !== user._id.toString()){
+            throw new ApiError(400,`user don't have permission to update this video`)
+        }  
+        // deleting the video
         const deletedVideo = await Video.findByIdAndDelete(videoId);
         if (!deletedVideo) {
             throw new ApiError(200, "there is an error while deleting the video")
@@ -152,12 +207,19 @@ const togglePublishStatus = asyncHandler(async (req, res) => {
         if (!mongoose.Types.ObjectId.isValid(videoId)) {
             return res.status(400).json({ error: 'Invalid videoId' });
         }
-        const isPublished = await Video.findById(videoId).select("isPublished");
+        // const user = req.user;
+        // if(!user){
+        //     throw new ApiError(400,"user not found!")
+        // }
+        const currentVideo = await Video.findById(videoId);
+        // if(currentVideo.owner.toString() !== user._id.toString()){
+        //     throw new ApiError(400,`user don't have permission to update this video`)
+        // }  
         const video = await Video.findByIdAndUpdate(
             videoId,
             {
                 $set: {
-                    isPublished: !isPublished.isPublished
+                    isPublished: !currentVideo.isPublished
                 }
             },
             {
